chore(client): rename Maps folder by MtvRoomConstraintsDetailsOnMap
