+++ packages/client/components/MtvRoomConstraintsDetailsOnMap/contract.ts
